requirements.txt
streamlit
langchain
langchain_groq
langchain_openai
langchain_community
pypdf
langchain_google_genai
faiss-cpu
python-dotenv
PyPDF2